Improve magic number error detection
Modify the PCAP schema to have better error messages regarding invalid magic
number, which is used to determine byteOrder. Previously, if the magic number
was invalid you would get an error message about invalid property value because
the byteOrder property was set to an error message. This was because the error
checking was performed in setVariable. Until support for an error() function or
something similar exists, split the checking and assignment into two
statements, one assert and one setVariable. Additionally, modify MagicNumber so
that it is hexBinary instead of uint32, making it much easier to quickly read
error messages.