Fix infosets to use correct namespacing
These infosets were extracted from the TDML test files, which had
incorrect namespaces on the elements. The TDML currently ignores
namespaces, so things work fine when run via the TDML Runner. But if you
try to unparse using the CLI, it fails. This fixes the namespaces to be
what is expected.